+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## v3.0.2 - 2023-09-18
+
+### Changed
+
+- Name of `generator` module to `gen`
+  - `generator` remains as an alias to `gen` for backward compatibility
+
+### Maintenance
+
+- Improved documentation about generator's clock rollback behavior
+
 ## v3.0.1 - 2023-07-29
 
 Most notably, v3 switches the letter case of generated IDs from uppercase (e.g.,